Date management major refactor

+++ Comets/OrbitViewer/ATimeSpan.cs
@@ -10,13 +10,13 @@ namespace Comets.OrbitViewer
 		public int Day { get; private set; }
 		public int Hour { get; private set; }
 		public int Minute { get; private set; }
-		public double Second { get; private set; }
+		public int Second { get; private set; }
 
 		#endregion
 
 		#region Constructor
 
-		public ATimeSpan(int year, int month, int day, int hour, int min, double sec)
+		public ATimeSpan(int year, int month, int day, int hour, int min, int sec)
 		{
 			this.Year = year;
 			this.Month = month;
